+++ packages/commonwealth/client/scripts/views/footer.tsx
@@ -6,14 +6,27 @@ import 'footer.scss';
 
 import { isNotUndefined } from '../helpers/typeGuards';
 
-type FooterAttrs = {
-  list: { text: string; redirectTo?: string; externalLink?: string }[];
-};
+const footercontents = [
+  { text: 'Blog', externalLink: 'https://blog.commonwealth.im' },
+  {
+    text: 'Jobs',
+    externalLink: 'https://angel.co/company/commonwealth-labs/jobs',
+  },
+  { text: 'Terms', redirectTo: '/terms' },
+  { text: 'Privacy', redirectTo: '/privacy' },
+  { text: 'Docs', externalLink: 'https://docs.commonwealth.im' },
+  {
+    text: 'Discord',
+    externalLink: 'https://discord.gg/t9XscHdZrG',
+  },
+  {
+    text: 'Telegram',
+    externalLink: 'https://t.me/HiCommonwealth',
+  },
+];
 
-export class Footer implements m.ClassComponent<FooterAttrs> {
-  view(vnode) {
-    const { list } = vnode.attrs;
-
+export class Footer implements m.ClassComponent {
+  view() {
     const redirectClick = (route) => {
       m.route.set(route);
     };
@@ -22,7 +35,7 @@ export class Footer implements m.ClassComponent<FooterAttrs> {
       <div class="Footer">
         <img src="/static/img/logo.svg" alt="Commonwealth" />
         <div class="footer-links-container">
-          {list.map((item) => {
+          {footercontents.map((item) => {
             return isNotUndefined(item.redirectTo) ? (
               <a
                 class="footer-link"
