+++ packages/commonwealth/client/styles/components/sidebar/index.scss
@@ -1,8 +1,5 @@
 @import 'shared';
 
-// TODO: Add SCSS variables to ensure that widths and heights of various
-// offsets and elements are uniform
-
 .Sidebar {
   background-color: $neutral-25;
   border-right: 1px solid $neutral-200;
@@ -11,19 +8,10 @@
   height: 100%;
   min-width: $sidebar-width;
   overflow-y: auto;
-  position: relative;
-  top: calc(#{$sidebar-community-header-height} - 1px);
   width: $sidebar-width;
 
   @include smallInclusive {
-    min-width: calc(100% - #{$quick-switcher-width});
-  }
-
-  .section-header {
-    color: $neutral-500;
-    font-size: 18px;
-    margin-bottom: 6px;
-    margin-top: 20px;
+    width: 100%;
   }
 
   .buttons-container {
@@ -32,7 +20,7 @@
     .subscription-button {
       display: flex;
       justify-content: center;
-      margin-bottom: 15px;
+      margin-bottom: 16px;
     }
   }
 }
