+++ packages/commonwealth/client/styles/components/sidebar/sidebar_quick_switcher.scss
@@ -10,35 +10,23 @@
   overflow: hidden;
   width: $quick-switcher-width;
 
-  &.chain-scoped {
-    padding: calc(#{$sidebar-community-header-height} + 4px) 4px 4px;
-  }
-
-  &.global {
-    padding: 4px;
-  }
-
   .community-nav-bar {
     align-items: center;
-    justify-content: center;
-    gap: 12px;
     display: flex;
     flex-direction: column;
-    justify-content: flex-start;
-    padding: 12px;
-    border-bottom: 1px solid $neutral-200;
-    width: min-content;
+    gap: 8px;
+    justify-content: center;
+    padding: 12px 0;
+    width: 100%;
   }
 
   .scrollable-community-bar {
     align-items: center;
     display: flex;
     flex-direction: column;
+    gap: 16px;
     overflow-y: auto;
-    padding: 12px 8px;
-
-    .SidebarQuickSwitcherItem {
-      margin-bottom: 16px;
-    }
+    padding: 12px 0;
+    width: 100%;
   }
 }
