+++ packages/commonwealth/client/styles/sublayout.scss
@@ -14,7 +14,7 @@
 
     .header-container {
       align-items: center;
-      border-bottom: 1px solid $neutral-50;
+      border-bottom: 1px solid $neutral-200;
       display: flex;
       flex-direction: row;
       height: $global-header-height;
@@ -32,6 +32,7 @@
 
     .sidebar-container {
       display: flex;
+      flex-direction: column;
 
       @include mediumSmallInclusive {
         display: none;
@@ -42,6 +43,12 @@
       }
 
       @include smallInclusive {
+        width: 100%;
+      }
+
+      .sidebar-inner-container {
+        display: flex;
+        height: 100%;
         width: 100%;
       }
     }
